Making application module executable. Organizing startup workarounds and fixes.
