+++ SickBeard.py
@@ -1,15 +1,7 @@
 #!/usr/bin/env python2.7
 # -*- coding: utf-8 -*
 """Script for backwards compatibility."""
-
-from __future__ import unicode_literals
-
-import sys
-
-from start import Application
-
+from medusa.__main__ import main
 
 if __name__ == '__main__':
-    # start application
-    application = Application()
-    application.start(sys.argv[1:])
+    main()
